snowflake_cortex/semantic_models/loader.py
"""Loader for semantic models from Snowflake."""

import logging
from typing import Dict, Any, Optional
import yaml
import snowflake.connector
from shared.config.settings import settings
from shared.utils.exceptions import SnowflakeCortexError

logger = logging.getLogger(__name__)


class SemanticModelLoader:
    """Loads semantic model YAML files from Snowflake."""
    
    def __init__(self):
        """Initialize the semantic model loader."""
        self.snowflake_config = settings.snowflake
        self.cache: Dict[str, Dict[str, Any]] = {}
        logger.info("Initialized Semantic Model Loader")
    
    async def load_semantic_model(
        self,
        model_name: Optional[str] = None
    ) -> Optional[Dict[str, Any]]:
        """
        Load a semantic model from Snowflake.
        
        Args:
            model_name: Optional name of the semantic model to load
        
        Returns:
            Semantic model dictionary or None if not found
        
        Raises:
            SnowflakeCortexError: If loading fails
        """
        try:
            # Use default model if not specified
            model_name = model_name or "default_semantic_model"
            
            # Check cache
            if model_name in self.cache:
                logger.debug(f"Returning cached semantic model: {model_name}")
                return self.cache[model_name]
            
            logger.info(f"Loading semantic model from Snowflake: {model_name}")
            
            # In production, this would load the semantic model YAML from Snowflake
            # The semantic model could be stored in:
            # 1. Snowflake stage
            # 2. Snowflake table
            # 3. External location (S3, etc.)
            
            # Example implementation:
            # conn = snowflake.connector.connect(...)
            # cursor = conn.cursor()
            # cursor.execute(
            #     f"SELECT content FROM semantic_models WHERE name = '{model_name}'"
            # )
            # result = cursor.fetchone()
            # if result:
            #     model_yaml = result[0]
            #     model_dict = yaml.safe_load(model_yaml)
            #     self.cache[model_name] = model_dict
            #     return model_dict
            
            # Placeholder semantic model
            semantic_model = {
                "name": model_name,
                "version": "1.0",
                "tables": [
                    {
                        "name": "example_table",
                        "description": "Example table for semantic queries",
                        "columns": [
                            {"name": "id", "type": "INTEGER", "description": "Unique identifier"},
                            {"name": "name", "type": "VARCHAR", "description": "Name field"},
                        ]
                    }
                ],
                "relationships": [],
                "common_queries": []
            }
            
            self.cache[model_name] = semantic_model
            logger.debug(f"Loaded semantic model: {model_name}")
            
            return semantic_model
            
        except Exception as e:
            logger.error(f"Error loading semantic model: {str(e)}")
            raise SnowflakeCortexError(f"Failed to load semantic model: {str(e)}") from e
    
    def clear_cache(self, model_name: Optional[str] = None):
        """
        Clear semantic model cache.
        
        Args:
            model_name: Optional specific model to clear, or all if None
        """
        if model_name:
            if model_name in self.cache:
                del self.cache[model_name]
                logger.debug(f"Cleared cache for semantic model: {model_name}")
        else:
            self.cache.clear()
            logger.debug("Cleared all semantic model cache")

